Refactor the utils library into a common folder that can be used by both tests and benchmarks
- Removes the sycl_blas dependency from clblas benchmarks (reduces compilation time of clblas benchmarks)
- Make a common library of utilities that both tests and benchmarks can use
