+++ benchmark/clblast/main.cpp
@@ -1,5 +1,5 @@
 #include "utils.hpp"
-#include "utils/extract_vendor_type.hpp"
+#include <common/extract_vendor_type.hpp>"
 
 int main(int argc, char** argv) {
   // Read the command-line arguments
