+++ common/include/common/extract_vendor_type.hpp
