+++ common/include/common/float_comparison.hpp
