+++ test/blas_test.hpp
@@ -39,12 +39,13 @@
 
 #include <sycl_blas.h>
 
+#include <common/cli_device_selector.hpp>
+#include <common/float_comparison.hpp>
+#include <common/print_queue_information.hpp>
+#include <common/system_reference_blas.hpp>
+
+#include <utils/quantization.hpp>
 #include "blas_test_macros.hpp"
-#include "utils/cli_device_selector.hpp"
-#include "utils/float_comparison.hpp"
-#include "utils/print_queue_information.hpp"
-#include "utils/quantization.hpp"
-#include "utils/system_reference_blas.hpp"
 
 struct Args {
   std::string program_name;
